+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/InheritedClassMap/ChildOfLineItemSource.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\InheritedClassMap;
+
+use Symfony\Component\ObjectMapper\Tests\Fixtures\NestedCollectionMapping\LineItemSource;
+
+class ChildOfLineItemSource extends LineItemSource
+{
+}