+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/InheritedClassMap/InheritedReverseChildSource.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\InheritedClassMap;
+
+class InheritedReverseChildSource extends InheritedReverseSource
+{
+}