+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/InheritedClassMap/InheritedReverseSource.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\InheritedClassMap;
+
+class InheritedReverseSource
+{
+    public function __construct(
+        public string $id = 'foo',
+    ) {
+    }
+}