+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/InheritedClassMap/InheritedReverseTarget.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\InheritedClassMap;
+
+class InheritedReverseTarget
+{
+    public string $id;
+}